requirement.txt
tensorboard==1.11.0
tensorflow-gpu==1.11.0
mpi4py==3.0.0
autograd==1.2
mujoco-py==1.50.1.65
-e git+git@github.com:williamd4112/gym.git@dabf2bf548d6d87b7c1109ba3bff14bb43c52aa6#egg=gym
absl-py==0.6.0
astor==0.7.1
atari-py==0.1.6
azure==1.0.3
azure-common==1.1.16
azure-mgmt==0.20.2
azure-mgmt-common==0.20.0
azure-mgmt-compute==0.20.1
azure-mgmt-network==0.20.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==0.20.1
azure-mgmt-storage==0.20.0
azure-nspkg==3.0.2
azure-servicebus==0.20.133333333333
azure-servicemanagement-legacy==0.20.2
azure-storage==0.20.3
cached-property==1.5.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
dill==0.2.8.2
future==0.16.0
gast==0.2.0
glfw==1.7.0
grpcio==1.16.0
h5py==2.8.0
idna==2.7
imageio==2.4.1
importlib-metadata==0.7
joblib==0.12.5
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lockfile==0.12.2
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.0
opencv-python==3.4.3.18
pandas==0.23.4
path.py==11.5.0
Pillow==5.3.0
progressbar2==3.38.0
protobuf==3.6.1
pycparser==2.19
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.2.2
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.6
pyzmq==17.1.2
requests==2.20.0
-e git+git@github.com:rll/rllab.git@ba78e4c16dc492982e648f117875b22af3965579#egg=rllab
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
termcolor==1.1.0
Theano==1.0.3
torchfile==0.1.0
tornado==5.1.1
tqdm==4.28.1
urllib3==1.24
-e git+https://github.com/facebookresearch/visdom.git@49b0e8f909a0570fae6aab125209e42e0b6c93b1#egg=visdom
websocket-client==0.53.0
Werkzeug==0.14.1
zmq==0.0.0
-e mbl_part/